#56 fix: removed duplicate handlClickOutsied
